# src/app/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mohammadiya App API",
    "version": "1.0.0",
    "description": "API for Admin, Teachers, Students"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Local API"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken"
      }
    },
    "schemas": {
...
      "ClassLevel": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "TimeSlot": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "label": { "type": "string" },
          "startTime": { "type": "string" },
          "endTime": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Student": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "studentName": { "type": "string" },
          "rollNumber": { "type": "string" },
          "grNumber": { "type": "string" },
          "gender": { "type": "string" },
          "age": { "type": "integer" },
          "fatherName": { "type": "string" },
          "mobile": { "type": "string" },
          "formNo": { "type": "string" },
          "admissionFee": { "type": "number" },
          "monthlyFees": { "type": "number" },
          "remarks": { "type": "string" },
          "hafizCategory": { "type": "string" },
          "fullTimeSubCategory": { "type": "string" },
          "classSession": { "$ref": "#/components/schemas/ClassSession" }
        }
      },
      "ClassSession": {
        "type": "object",
        "properties": {
             "id": { "type": "string" },
             "classLevel": { "$ref": "#/components/schemas/ClassLevel" },
             "timeSlot": { "$ref": "#/components/schemas/TimeSlot" }
        }
      },
...
    }
  },
  "paths": {
...
    "/payments": {
      "get": {
        "summary": "List Payments / Search Students",
        "tags": ["Payments"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Search by Name or Roll Number" },
          { "name": "classId", "in": "query", "schema": { "type": "string" } },
          { "name": "timeSlotId", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["PAID", "UNPAID", "ALL"] } }
        ],
        "responses": {
          "200": {
            "description": "List of students with payment status",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Student" } }, "pagination": { "type": "object" } } } } }
          }
        }
      },
      "post": {
        "summary": "Record Payment (PayNow / Advance)",
        "tags": ["Payments"],
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["studentId", "amount", "months"],
                "properties": {
                  "studentId": { "type": "string" },
                  "amount": { "type": "number" },
                  "months": { "type": "array", "items": { "type": "string" }, "description": "Array of YYYY-MM" },
                  "remarks": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Payment recorded" }
        }
      }
    },
    "/payments/receipt": {
      "get": {
        "summary": "Get Receipt Details",
        "tags": ["Payments"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          { "name": "paymentId", "in": "query", "schema": { "type": "string" } },
          { "name": "studentId", "in": "query", "schema": { "type": "string" }, "description": "Get latest receipt for student" }
        ],
        "responses": {
          "200": {
            "description": "Receipt details",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/dashboard/admissionapplications": {
      "get": {
        "summary": "List Admission Applications",
        "tags": ["Dashboard"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "List of applications" }
        }
      },
      "put": {
        "summary": "Update Admission Status",
        "tags": ["Dashboard"],
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["studentId", "status"],
                "properties": {
                  "studentId": { "type": "string" },
                  "status": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Status updated" }
        }
      }
    },
...
    "/dashboard/overview": {
      "get": {
        "summary": "Get Dashboard Overview Stats",
        "tags": ["Dashboard"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          { "name": "month", "in": "query", "schema": { "type": "integer" } },
          { "name": "year", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Overview stats" }
        }
      }
    }
  }
}